requirements.txt

# Based on Raspberry Pi OS (Bullseye) 
Python 3.9

Python-vlc==3.0.*